tests/types/test_gridworks_event_snapshot_spaceheat.py
"""Tests gridworks.event.snapshot.spaceheat type, version 000"""

import json

import pytest
from gw.errors import GwTypeError
from pydantic import ValidationError

from gwp.types import GridworksEventSnapshotSpaceheat
from gwp.types import GridworksEventSnapshotSpaceheat_Maker as Maker


def test_gridworks_event_snapshot_spaceheat_generated() -> None:
    t = GridworksEventSnapshotSpaceheat(
        message_id=,
        time_n_s=1699886100019488593,
        src='hw1.isone.me.versant.keene.beech.scada',
        snap=,
    )

    d = {
        "MessageId": ,
        "TimeNS": 1699886100019488593,
        "Src": 'hw1.isone.me.versant.keene.beech.scada',
        "Snap": ,
        "TypeName": "gridworks.event.snapshot.spaceheat",
        "Version": "000",
    }

    assert t.as_dict() == d

    with pytest.raises(GwTypeError):
        Maker.type_to_tuple(d)

    with pytest.raises(GwTypeError):
        Maker.type_to_tuple('"not a dict"')

    # Test type_to_tuple
    gtype = json.dumps(d)
    gtuple = Maker.type_to_tuple(gtype)
    assert gtuple == t

    # test type_to_tuple and tuple_to_type maps
    assert Maker.type_to_tuple(Maker.tuple_to_type(gtuple)) == gtuple

    ######################################
    # GwTypeError raised if missing a required attribute
    ######################################

    d2 = dict(d)
    del d2["TypeName"]
    with pytest.raises(GwTypeError):
        Maker.dict_to_tuple(d2)

    d2 = dict(d)
    del d2["MessageId"]
    with pytest.raises(GwTypeError):
        Maker.dict_to_tuple(d2)

    d2 = dict(d)
    del d2["TimeNS"]
    with pytest.raises(GwTypeError):
        Maker.dict_to_tuple(d2)

    d2 = dict(d)
    del d2["Src"]
    with pytest.raises(GwTypeError):
        Maker.dict_to_tuple(d2)

    d2 = dict(d)
    del d2["Snap"]
    with pytest.raises(GwTypeError):
        Maker.dict_to_tuple(d2)

    ######################################
    # Behavior on incorrect types
    ######################################

    d2 = dict(d, TimeNS="1699886100019488593.1")
    with pytest.raises(ValidationError):
        Maker.dict_to_tuple(d2)

    ######################################
    # ValidationError raised if TypeName is incorrect
    ######################################

    d2 = dict(d, TypeName="not the type name")
    with pytest.raises(ValidationError):
        Maker.dict_to_tuple(d2)

    ######################################
    # ValidationError raised if primitive attributes do not have appropriate property_format
    ######################################

    d2 = dict(d, MessageId="d4be12d5-33ba-4f1f-b9e5")
    with pytest.raises(ValidationError):
        Maker.dict_to_tuple(d2)

    d2 = dict(d, Src="a.b-h")
    with pytest.raises(ValidationError):
        Maker.dict_to_tuple(d2)
